pqcrypto_dilithium

Module dilithium5

dilithium5

These bindings use the clean version from PQClean

Example

// if using pqcrypto-dilithium
use pqcrypto_dilithium::dilithium5::*;
// or if using the pqcrypto crate:
// use pqcrypto::sign::dilithium5::*;
let message = vec![0, 1, 2, 3, 4, 5];
let (pk, sk) = keypair();
let sm = sign(&message, &sk);
let verifiedmsg = open(&sm, &pk).unwrap();
assert!(verifiedmsg == message);

Macros

detached_signature 🔒

gen_keypair 🔒

gen_signature 🔒

open_signed 🔒

simple_struct 🔒

verify_detached_sig 🔒

Structs

DetachedSignature

PublicKey

SecretKey

SignedMessage

Functions

detached_sign

Create a detached signature on the message

keypair

Generate a dilithium5 keypair

open

Open the signed message and if verification succeeds return the message

public_key_bytes

Get the number of bytes for a public key

secret_key_bytes

Get the number of bytes for a secret key

sign

Sign the message and return the signed message.

signature_bytes

Get the number of bytes that a signature occupies

verify_detached_signature

Verify the detached signature